+++ src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraph.const.tsx
@@ -1,5 +1,6 @@
 export const FORECASTS_GRAPH_HEIGHT = 56;
 export const FORECASTS_GRAPH_BOTTOM_OFFSET = 20;
+export const FORECASTS_GRAPH_BOTTOM_PADDING = 8;
 export const FORECASTS_GRAPH_MAX_VALUE = 100;
 export const FORECASTS_GRAPH_SPACING = 48;
 export const FORECASTS_GRAPH_POINT_SIZE = 6;
